profile_details/requirements.txt
asgiref==3.8.1
boto3==1.34.117
botocore==1.34.131
Django==4.2.13
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
jmespath==1.0.1
mysqlclient==2.2.4
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
urllib3==1.26.19
django-import-export
dj-database-url
psycopg2
